Make auth settings async
