+++ azure-mgmt-appservice/src/main/java/com/microsoft/azure/management/appservice/WebAppAuthentication.java
@@ -277,113 +277,4 @@ public interface WebAppAuthentication extends
                 WithExternalRedirectUrls<ParentT> {
         }
     }
-
-    /**
-     * Grouping of all the web app authentication update stages.
-     */
-    interface UpdateStages {
-        /**
-         * A web app authentication update allowing the default authentication provider to be set.
-         * @param <ParentT> the return type of the final {@link Update#parent()}
-         */
-        interface WithDefaultAuthenticationProvider<ParentT> {
-            /**
-             * Does not require login by default.
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withAnonymousAuthentication();
-
-            /**
-             * Specifies the default authentication provider.
-             * @param provider the default authentication provider
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withDefaultAuthenticationProvider(BuiltInAuthenticationProvider provider);
-        }
-
-        /**
-         * A web app authentication update allowing detailed provider information to be specified.
-         * @param <ParentT> the return type of the final {@link Update#parent()}
-         */
-        interface WithAuthenticationProvider<ParentT> {
-            /**
-             * Specifies the provider to be Active Directory and its client ID and issuer URL.
-             * @param clientId the AAD app's client ID
-             * @param issuerUrl the token issuer URL in the format of https://sts.windows.net/(tenantId)
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withActiveDirectory(String clientId, String issuerUrl);
-
-            /**
-             * Specifies the provider to be Facebook and its app ID and app secret.
-             * @param appId the Facebook app ID
-             * @param appSecret the Facebook app secret
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withFacebook(String appId, String appSecret);
-
-            /**
-             * Specifies the provider to be Google and its client ID and client secret.
-             * @param clientId the Google app's client ID
-             * @param clientSecret the Google app's client secret
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withGoogle(String clientId, String clientSecret);
-
-            /**
-             * Specifies the provider to be Twitter and its API key and API secret.
-             * @param apiKey the Twitter app's API key
-             * @param apiSecret the Twitter app's API secret
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withTwitter(String apiKey, String apiSecret);
-
-            /**
-             * Specifies the provider to be Microsoft and its client ID and client secret.
-             * @param clientId the Microsoft app's client ID
-             * @param clientSecret the Microsoft app's client secret
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withMicrosoft(String clientId, String clientSecret);
-        }
-
-        /**
-         * A web app authentication update allowing token store to be specified.
-         * @param <ParentT> the return type of the final {@link Update#parent()}
-         */
-        interface WithTokenStore<ParentT> {
-            /**
-             * Specifies if token store should be enabled.
-             * @param enabled true if token store should be enabled
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withTokenStore(boolean enabled);
-        }
-
-        /**
-         * A web app authentication update allowing branch to be specified.
-         * @param <ParentT> the return type of the final {@link Update#parent()}
-         */
-        interface WithExternalRedirectUrls<ParentT> {
-            /**
-             * Adds an external redirect URL.
-             * @param url the external redirect URL
-             * @return the next stage of the web app authentication update
-             */
-            Update<ParentT> withAllowedExternalRedirectUrl(String url);
-        }
-    }
-
-    /**
-     * The entirety of a web app authentication update as part of a web app update.
-     * @param <ParentT> the return type of the final {@link Update#parent()}
-     */
-    interface Update<ParentT> extends
-            Settable<ParentT>,
-            UpdateStages.WithDefaultAuthenticationProvider<ParentT>,
-            UpdateStages.WithAuthenticationProvider<ParentT>,
-            UpdateStages.WithTokenStore<ParentT>,
-            UpdateStages.WithExternalRedirectUrls<ParentT> {
-    }
-
 }
