+++ azure-mgmt-appservice/src/main/java/com/microsoft/azure/management/appservice/implementation/AppServiceBaseImpl.java
@@ -141,50 +141,6 @@ abstract class AppServiceBaseImpl<
     }
 
     @Override
-    public Map<String, AppSetting> getAppSettings() {
-        return getAppSettingsAsync().toBlocking().single();
-    }
-
-    public Observable<Map<String, AppSetting>> getAppSettingsAsync() {
-        return Observable.zip(listAppSettings(), listSlotConfigurations(), new Func2<StringDictionaryInner, SlotConfigNamesResourceInner, Map<String, AppSetting>>() {
-            @Override
-            public Map<String, AppSetting> call(final StringDictionaryInner appSettingsInner, final SlotConfigNamesResourceInner slotConfigs) {
-                if (appSettingsInner != null && appSettingsInner.properties() != null) {
-                    return Maps.asMap(appSettingsInner.properties().keySet(), new Function<String, AppSetting>() {
-                        @Override
-                        public AppSetting apply(String input) {
-                            return new AppSettingImpl(input, appSettingsInner.properties().get(input),
-                                    slotConfigs.appSettingNames() != null && slotConfigs.appSettingNames().contains(input));
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    @Override
-    public Map<String, ConnectionString> getConnectionStrings() {
-        return getConnectionStringsAsync().toBlocking().single();
-    }
-
-    public Observable<Map<String, ConnectionString>> getConnectionStringsAsync() {
-        return Observable.zip(listConnectionStrings(), listSlotConfigurations(), new Func2<ConnectionStringDictionaryInner, SlotConfigNamesResourceInner, Map<String, ConnectionString>>() {
-            @Override
-            public Map<String, ConnectionString> call(final ConnectionStringDictionaryInner connectionStringsInner, final SlotConfigNamesResourceInner slotConfigs) {
-                if (connectionStringsInner != null && connectionStringsInner.properties() != null) {
-                    return Maps.asMap(connectionStringsInner.properties().keySet(), new Function<String, ConnectionString>() {
-                        @Override
-                        public ConnectionString apply(String input) {
-                            return new ConnectionStringImpl(input, connectionStringsInner.properties().get(input),
-                                    slotConfigs.connectionStringNames() != null && slotConfigs.connectionStringNames().contains(input));
-                        }
-                    });
-                }
-            }
-        });
-    }
-
-    @Override
     @SuppressWarnings("unchecked")
     public Observable<Map<String, HostNameBinding>> getHostNameBindingsAsync() {
         return this.manager().inner().webApps().listHostNameBindingsAsync(resourceGroupName(), name())
